Move types to more appropriate files
