+++ cli/src/cli.rs
@@ -39,13 +39,13 @@ use solana_sdk::{
     bpf_loader,
     clock::{Epoch, Slot, DEFAULT_TICKS_PER_SECOND},
     commitment_config::CommitmentConfig,
+    decode_error::DecodeError,
     fee_calculator::FeeCalculator,
     hash::Hash,
     instruction::InstructionError,
     loader_instruction,
     message::Message,
     native_token::lamports_to_sol,
-    program_utils::DecodeError,
     pubkey::{Pubkey, MAX_SEED_LEN},
     signature::{Keypair, Signature, Signer, SignerError},
     signers::Signers,
