+++ programs/move_loader/src/processor.rs
@@ -6,13 +6,13 @@ use log::*;
 use num_derive::{FromPrimitive, ToPrimitive};
 use serde_derive::{Deserialize, Serialize};
 use solana_sdk::{
-    account::KeyedAccount,
+    account::{is_executable, next_keyed_account, KeyedAccount},
     account_utils::State,
+    decode_error::DecodeError,
     entrypoint_native::InvokeContext,
     instruction::InstructionError,
     move_loader::id,
-    program_utils::DecodeError,
-    program_utils::{is_executable, limited_deserialize, next_keyed_account},
+    program_utils::limited_deserialize,
     pubkey::Pubkey,
 };
 use thiserror::Error;
