+++ programs/stake/src/stake_instruction.rs
@@ -6,10 +6,11 @@ use log::*;
 use num_derive::{FromPrimitive, ToPrimitive};
 use serde_derive::{Deserialize, Serialize};
 use solana_sdk::{
-    account::{get_signers, KeyedAccount},
+    account::{get_signers, next_keyed_account, KeyedAccount},
     clock::{Epoch, UnixTimestamp},
+    decode_error::DecodeError,
     instruction::{AccountMeta, Instruction, InstructionError},
-    program_utils::{limited_deserialize, next_keyed_account, DecodeError},
+    program_utils::limited_deserialize,
     pubkey::Pubkey,
     system_instruction,
     sysvar::{self, clock::Clock, rent::Rent, stake_history::StakeHistory, Sysvar},
