+++ programs/vest/src/vest_processor.rs
@@ -7,9 +7,9 @@ use chrono::prelude::*;
 use solana_config_program::date_instruction::DateConfig;
 use solana_config_program::get_config_data;
 use solana_sdk::{
-    account::{Account, KeyedAccount},
+    account::{next_keyed_account, Account, KeyedAccount},
     instruction::InstructionError,
-    program_utils::{limited_deserialize, next_keyed_account},
+    program_utils::limited_deserialize,
     pubkey::Pubkey,
 };
 use std::cell::RefMut;
