+++ programs/vote/src/vote_instruction.rs
@@ -10,10 +10,11 @@ use num_derive::{FromPrimitive, ToPrimitive};
 use serde_derive::{Deserialize, Serialize};
 use solana_metrics::inc_new_counter_info;
 use solana_sdk::{
-    account::{get_signers, KeyedAccount},
+    account::{get_signers, next_keyed_account, KeyedAccount},
+    decode_error::DecodeError,
     hash::Hash,
     instruction::{AccountMeta, Instruction, InstructionError},
-    program_utils::{limited_deserialize, next_keyed_account, DecodeError},
+    program_utils::limited_deserialize,
     pubkey::Pubkey,
     system_instruction,
     sysvar::{self, clock::Clock, slot_hashes::SlotHashes, Sysvar},
