+++ runtime/src/legacy_system_instruction_processor0.rs
@@ -1,9 +1,9 @@
 use log::*;
 use solana_sdk::{
-    account::{get_signers, Account, KeyedAccount},
+    account::{get_signers, next_keyed_account, Account, KeyedAccount},
     instruction::InstructionError,
     nonce::Account as NonceAccount,
-    program_utils::{limited_deserialize, next_keyed_account},
+    program_utils::limited_deserialize,
     pubkey::Pubkey,
     system_instruction::{SystemError, SystemInstruction, MAX_PERMITTED_DATA_LENGTH},
     system_program,
