+++ runtime/src/system_instruction_processor.rs
@@ -1,10 +1,10 @@
 use log::*;
 use solana_sdk::{
-    account::{get_signers, Account, KeyedAccount},
+    account::{get_signers, next_keyed_account, Account, KeyedAccount},
     account_utils::StateMut,
     instruction::InstructionError,
     nonce::{self, Account as NonceAccount},
-    program_utils::{limited_deserialize, next_keyed_account},
+    program_utils::limited_deserialize,
     pubkey::Pubkey,
     system_instruction::{SystemError, SystemInstruction, MAX_PERMITTED_DATA_LENGTH},
     system_program,
