+++ sdk/src/lib.rs
@@ -13,6 +13,7 @@ pub mod account_utils;
 pub mod bpf_loader;
 pub mod clock;
 pub mod commitment_config;
+pub mod decode_error;
 pub mod entrypoint_native;
 pub mod epoch_info;
 pub mod epoch_schedule;
