+++ sdk/src/system_instruction.rs
@@ -1,7 +1,7 @@
 use crate::{
+    decode_error::DecodeError,
     instruction::{AccountMeta, Instruction},
     nonce,
-    program_utils::DecodeError,
     pubkey::Pubkey,
     system_program,
     sysvar::{recent_blockhashes, rent},
